server/requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
langchain-core==0.2.8
langchain-text-splitters==0.2.1
faster-whisper==1.0.2
matplotlib==3.9.0
pydub==0.25.1
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.7.0/fr_core_news_lg-3.7.0-py3-none-any.whl#sha256=c2d98eb9073ff0f3d865179448e7f1868f27b2643c0e7215acc2438cd9eae554
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl#sha256=37e9f1f6a278a5138fabdabcc92cc559da917f9b24c76f0adf6758720d7eab10
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
pytube==15.0.0

nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105


# torch==2.3.1
# torchvision==0.18.1
watchdog==4.0.1
watchfiles==0.22.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
unstructured==0.14.8
unstructured-client==0.23.7
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
pillow_heif==0.16.0
einops==0.8.0
flash-attn==2.5.9.post1
transformers==4.41.2
sentencepiece==0.2.0
pytesseract==0.3.10
unstructured.pytesseract==0.3.12
sacremoses==0.1.1
surya-ocr==0.4.14
google-api-python-client==2.136.0
boto3==1.34.131
fastapi-cache2==0.2.1
aioredis==1.3.1
hiredis==2.3.2
redis==4.6.0
psycopg==3.2.1
psycopg2-binary==2.9.9
selenium==4.27.1
webdriver-manager==4.0.2